+++ src/facades/index.js
@@ -1,0 +1,5 @@
+const TradingWalletFacade = require('./TradingWalletFacade')
+
+module.exports = {
+  TradingWalletFacade,
+}