+++ src/factories/Erc20TokenTransactionBuilder.js
@@ -87,6 +87,25 @@ class Erc20TokenTransactionBuilder extends BaseTransactionBuilder {
     )
     return transactionDraft
   }
+
+  buildGetBalanceOfTransactionDraft(personalWalletAddress) {
+    this.constructor.checkEtherumAddress(personalWalletAddress)
+
+    const smartContractMethodName = 'balanceOf'
+    const smartContractParams = [personalWalletAddress]
+    const transactionParams = {
+      from: personalWalletAddress,
+      to: this.erc20TokenSmartContractAddress,
+    }
+
+    const transactionDraft = this.transactionLib.buildDraft(this.erc20TokenSmartContractInstance,
+      transactionParams, smartContractMethodName, smartContractParams)
+    this.log.debug(
+      { fn: 'buildGetBalanceOfTransactionDraft', personalWalletAddress },
+      'Erc20 token get balance of transaction draft builded successfully.',
+    )
+    return transactionDraft
+  }
 }
 
 module.exports = Erc20TokenTransactionBuilder